# api/docs/swagger.yaml
definitions:
...
  models.CreateStaffTarif:
    properties:
      amount_for_card:
        type: number
      amount_for_cash:
        type: number
      name:
        type: string
      type:
        description: (fixed, percent)
        type: string
    type: object
...
  models.GetAllStaffTarif:
    properties:
      count:
        type: integer
      staff_tarifs:
        items:
          $ref: '#/definitions/models.StaffTarif'
        type: array
    type: object
...
  models.StaffTarif:
    properties:
      amount_for_card:
        type: number
      amount_for_cash:
        type: number
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      type:
        description: (fixed, percent)
        type: string
      updated_at:
        type: string
    type: object
...
  response.ErrorResp:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /tariff:
    get:
      consumes:
      - application/json
      description: gets all tariff based on limit, page and search by name
      parameters:
      - default: 10
        description: limit
        in: query
        minimum: 1
        name: limit
        type: integer
      - default: 1
        description: page
        in: query
        minimum: 1
        name: page
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAllStaffTarif'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: GET  ALL BRANCHS
      tags:
      - TARIFF
    post:
      consumes:
      - application/json
      description: CREATES TARIFF BASED ON GIVEN DATA
      parameters:
      - description: tariff data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CreateStaffTarif'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: CREATES TARIFF
      tags:
      - TARIFF
  /tariff/{id}:
    delete:
      consumes:
      - application/json
      description: DELETES TARIFF BASED ON ID
      parameters:
      - description: id of tariff
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: DELETE TARIFF BY ID
      tags:
      - TARIFF
    get:
      consumes:
      - application/json
      description: get tariff by ID
      parameters:
      - description: Tariff ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StaffTarif'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: GET BY ID
      tags:
      - TARIFF
    put:
      consumes:
      - application/json
      description: UPDATES TARIFF BASED ON GIVEN DATA AND ID
      parameters:
      - description: id of tariff
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: tariff data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CreateStaffTarif'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: UPDATE TARIFF BY ID
      tags:
      - TARIFF
...
swagger: "2.0"